+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Support dynamically changeable logging configuration of Kafka brokers
 * Support dynamically changeable logging configuration of Kafka MirrorMaker2
 * Updated Jaeger tracing libraries to support the W3C trace context headers
+* Add support for `client.rack` property for Kafka Connect to use `fetch from closest replica` feature. 
 
 ### Deprecations and removals